third_party/blink/web_tests/platform/generic/fast/dom/prototype-property-expected.txt

Make sure the 'prototype' property on generated Web IDL interfaces is { DontDelete | ReadOnly }.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tryToDeletePrototype(window.HTMLElement) is false
PASS tryToSetPrototype(window.HTMLElement, null) is false
PASS tryToSetPrototype(window.HTMLElement, undefined) is false
PASS tryToSetPrototype(window.HTMLElement, 1) is false
PASS tryToSetPrototype(window.HTMLElement, window.Object.prototype) is false
PASS successfullyParsed is true

TEST COMPLETE
